core: move InsertPoint to Rewriter
